requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
authlib==1.3.2
jinja2==3.1.4
python-multipart==0.0.12
itsdangerous==2.2.0

